Criada classe TestDouble: testar de igualdade em doubles com certa tolerância; Adicionado teste da window; Adicionado sobrecarga para comparação de doubles em testFunction; Viewport::area retorna um inteiro.

+++ view/viewport.h
@@ -17,12 +17,13 @@ struct Viewport {
      */
     static Viewport generateViewport( Drawable* );
 
-    double area() const;
+    /* Quantidade de pixels na viewport. */
+    int area() const;
 };
 
 //Implementação de area()
 
-inline double Viewport::area() const {
+inline int Viewport::area() const {
     return (xmax - xmin) * (ymax - ymin);
 }
